web-animations/animation-node-list/animation-node-list-length.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationNodeList length attribute test</title>
<meta name="assert" content="The number of animation nodes in the list">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationnodelist-length">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var nodes = [new AnimationGroup([]), new AnimationSequence([])];
    nodes.forEach(function(node) {
        assert_equals(node.children.length, 0, type(node) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns zero for empty AnimationNodeList');

test(function() {
    var node1 = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var node2 = new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION);
    var nodes = [new AnimationGroup([node1]), new AnimationSequence([node2])];
    nodes.forEach(function(node) {
        assert_equals(node.children.length, 1, type(node) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns 1 for AnimationNodeList with single element');

test(function() {
    var test = this;
    var parents = [AnimationGroup, AnimationSequence];
    parents.forEach(function(parentCtor) {
        var node1 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node2 = new AnimationGroup([]);
        var node3 = new AnimationSequence([]);
        var parent = new parentCtor([node1, node2, node3]);

        assert_equals(parent.children.length, 3, parentCtor.name + 'children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list');

test(function() {
    var test = this;
    var nodes = [new AnimationGroup([]), new AnimationSequence([])];
    nodes.forEach(function(node1) {
        assert_equals(node1.children.length, 0, type(node1) + '.children.length ' +
            'should return the number of animation nodes in the list');
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        node1.append(node2);
        assert_equals(node1.children.length, 1, type(node1) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationGroup.append()');

test(function() {
    var test = this;
    var nodes = [new AnimationGroup([]), new AnimationSequence([])];
    nodes.forEach(function(node1) {
        assert_equals(node1.children.length, 0, type(node1) + '.children.length ' +
            'should return the number of animation nodes in the list');
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        node1.prepend(node2);
        assert_equals(node1.children.length, 1, type(node1) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationGroup.prepend()');

test(function() {
    var test = this;
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION)
    ];
    var parents = [new AnimationGroup([nodes[0]]), new AnimationSequence([nodes[1]])];

    nodes.forEach(function(node1, index) {
        assert_equals(parents[index].children.length, 1, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        node1.before(node2);
        assert_equals(parents[index].children.length, 2, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationNode.before()');

test(function() {
    var test = this;
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION)
    ];
    var parents = [new AnimationGroup([nodes[0]]), new AnimationSequence([nodes[1]])];

    nodes.forEach(function(node1, index) {
        assert_equals(parents[index].children.length, 1, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        node1.after(node2);
        assert_equals(parents[index].children.length, 2, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationNode.after()');

test(function() {
    var test = this;
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION)
    ];
    var parents = [new AnimationGroup([nodes[0]]), new AnimationSequence([nodes[1]])];

    nodes.forEach(function(node1, index) {
        assert_equals(parents[index].children.length, 1, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node3 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);

        node1.replace(node2, node3);
        assert_equals(parents[index].children.length, 2, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationNode.replace()');

test(function() {
    var nodes = [
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION),
        new Animation(createDiv(this), KEYFRAMES, ANIMATION_DURATION)
    ];
    var parents = [new AnimationGroup([nodes[0]]), new AnimationSequence([nodes[1]])];

    nodes.forEach(function(node, index) {
        assert_equals(parents[index].children.length, 1, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
        node.remove();
        assert_equals(parents[index].children.length, 0, type(parents[index]) + '.children.length ' +
            'should return the number of animation nodes in the list');
    });
}, 'Test AnimationNodeList.length returns the number of animation nodes in the list, ' +
    'the list is changed by method AnimationNode.remove()');
</script>
</body>
